requirements.txt
flask
python-dotenv
pymongo
openai 